team_member_management/requirements.txt
Babel==2.7.0
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
gunicorn==19.9.0
phonenumbers==8.10.22
psycopg2-binary==2.8.4
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
